pyproject.toml
[project]
name = "kirok-mcp"
version = "1.0.0"
description = "Persistent memory for AI agents — Retain, Recall, Reflect via MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Tadahiko Fujikawa" },
]
keywords = [
    "mcp",
    "memory",
    "agent",
    "ai",
    "embeddings",
    "semantic-search",
    "gemini",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/TadFuji/kirok-mcp"
Repository = "https://github.com/TadFuji/kirok-mcp"
Issues = "https://github.com/TadFuji/kirok-mcp/issues"

[project.scripts]
kirok-mcp = "kirok_mcp.server:main"
kirok-doctor = "kirok_mcp.diagnostics:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kirok_mcp"]
